requirements.txt
fastapi[all]==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pydantic==2.6.4
redis==5.0.1
celery==5.3.6
pika==1.3.2
python-jose==3.3.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
requests==2.31.0
gunicorn==21.2.0
prometheus-client==0.20.0
